% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmocean.R
\docType{data}
\name{cmocean}
\alias{cmocean}
\title{cmocean colors
The cmocean color palette by Kristen Thyng}
\format{An object of class \code{list} of length 22.}
\usage{
cmocean
}
\description{
str(cmocean)
List of 22
$ algae
$ amp
$ balance
$ curl
$ deep
$ delta
$ dense
$ diff
$ gray
$ haline
$ ice
$ matter
$ oxy
$ phase
$ rain
$ solar
$ speed
$ tarn
$ tempo
$ topo
$ thermal
$ turbid
}
\keyword{datasets}
