% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataInfo.R
\docType{data}
\name{dataInfo}
\alias{dataInfo}
\title{dataInfo Data}
\format{An object of class \code{info} of length 3.}
\usage{
dataInfo
}
\description{
pre-Download of 'rerddap' info needed for examples
so can run within CRAN Time limits
}
\details{
obtained using
dataInfo <- rerddap::info('erdHadISST')
}
\keyword{datasets}
