% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFIleCoords.R
\name{getfileCoords}
\alias{getfileCoords}
\title{Get Coordinate  (Dimension) Data from ERDDAP Dataset}
\usage{
getfileCoords(datasetID, dataCoords, urlbase)
}
\arguments{
\item{urlbase}{A character string giving the base URL of the ERDDAP server}

\item{dataStruct}{A structure describing the dataset from erddapStruct.rda}
}
\value{
A list containing the values of the coordinate variables
}
\description{
Get Coordinate  (Dimension) Data from ERDDAP Dataset
}
\keyword{(dimension)}
\keyword{\code{getFileCoords}}
\keyword{an}
\keyword{coordinate}
\keyword{dataset}
\keyword{function}
\keyword{gets}
\keyword{internal}
\keyword{is}
\keyword{of}
\keyword{requested}
\keyword{that}
\keyword{the}
\keyword{variables}
