% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBsst.R
\docType{data}
\name{MBsst}
\alias{MBsst}
\title{MBsst Data}
\format{
An object of class \code{list} (inherits from \code{rxtracto3D}) of length 6.
}
\usage{
MBsst
}
\description{
pre-Download of Pacific West Coast SST fro use in `plotBBox()` example
can run within CRAN Time limits
}
\details{
obtained using the `rxtracto_3D()` command
dataInfo <- rerddap::info('erdMBsstd1day')
parameter <- 'sst'
xcoord <- c(230, 231)
ycoord <- c(33, 34)
tcoord <- c('2006-01-15', '2006-01-15')
zcoord <- c(0., 0.)
MBsst <- rxtracto_3D(dataInfo, parameter, xcoord = xcoord, ycoord = ycoord,
                     tcoord = tcoord, zcoord = zcoord)
}
\keyword{datasets}
