% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PB_Argos.R
\docType{data}
\name{PB_Argos}
\alias{PB_Argos}
\title{Polar Bear Track Data}
\format{
A data frame with 1919 obs. of 4 variables:
\describe{
  \item{DateTime}{time of observation}
  \item{QualClass}{Quality Flag}
  \item{Lat}{latitude of observation}
  \item{Lon}{longitude of observation}
  }
}
\source{
\describe{
  \item{dryad dataset}{https://datadryad.org/stash/dataset/doi:10.5061/dryad.4qrfj6q96}
  \item{Citation}{Auger-Méthé, Marie; Derocher, Andrew E. (2021). Argos and GPS data for a polar bear track [Dataset]. Dryad. https://doi.org/10.5061/dryad.4qrfj6q96}
  \item{Associated Paper}{Auger-Méthé, M.,  Newman, K.,  Cole, D.,  Empacher, F.,  Gryba, R.,  King, A. A.,  Leos-Barajas, V.,  Mills Flemming, J., Nielsen, A.,  Petris, G., and  Thomas, L..  2021.  A guide to state–space modeling of ecological time series. Ecological Monographs  91(4):e01470. 10.1002/ecm.1470}
}
}
\usage{
PB_Argos
}
\description{
Telemetry data of a tagged polar bear
}
\keyword{datasets}
