% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBox.R
\name{plotBox}
\alias{plotBox}
\title{plot result of xtracto_3D or rxtracto_3D}
\usage{
plotBox(resp, plotColor = "viridis", time = NA, animate = FALSE,
  name = NA, myFunc = NA, maxpixels = 10000)
}
\arguments{
\item{resp}{data frame returned from rxtracto() or xtracto()}

\item{plotColor}{the color to use in plot from rerddap}

\item{time}{a function to map multi-time to one,  or else identity for animation}

\item{animate}{if multiple times, if TRUE will animate the maps}

\item{myFunc}{function of one argument to transform the data}

\item{maxpixels}{maximum numbe rof pixels to use in making the map - controls resolution}
}
\value{
a plotdap plot
}
\description{
\code{plotBox} is a function to plot the results from
rxtracto() and xtracto()
}
\examples{
tagData <- Marlintag38606
xpos <- tagData$lon
ypos <- tagData$lat
tpos <- tagData$date
zpos <- rep(0., length(xpos))
urlbase <- 'http://upwell.pfeg.noaa.gov/erddap'
swchlInfo <- rerddap::info('erdSWchla8day')
swchl <- rxtracto(swchlInfo, parameter = 'chlorophyll', xcoord = xpos, ycoord = ypos, tcoord = tpos, zcoord = zpos, xlen = .2, ylen = .2)
plotBox(xpos, ypos, swchl, plotColor = 'chlorophyll')
}
