% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxtracto_3D.R
\name{rxtracto_3D}
\alias{rxtracto_3D}
\title{\code{rxtracto_3D} uses the R program rerddap to extract environmental data
from an ERDDAP server in an (x,y,z, time) bounding box.
The same call could be made directly in rerddap,
but function is maintained as it is used in the polygon routine.}
\usage{
rxtracto_3D(dataInfo, parameter = NULL, xcoord = NULL, ycoord = NULL,
  zcoord = NULL, tcoord = NULL, xName = "longitude", yName = "latitude",
  zName = "altitude", tName = "time",
  urlbase = "http://upwell.pfeg.noaa.gov/erddap", verbose = FALSE)
}
\arguments{
\item{dataInfo}{- the return from an rerddap "info" call to an ERDDAP server}

\item{parameter}{- character string containing the name of the parameter to extract}

\item{xcoord}{- a real array with the x-coordinates of the trajectory (if longitude in #'   decimal degrees East, either 0-360 or -180 to 180)}

\item{ycoord}{-  a real array with the y-coordinate of the trajectory (if latitude in
decimal degrees N; -90 to 90)}

\item{zcoord}{-  a real number with the z-coordinate (usually altitude or depth)}

\item{tcoord}{- a character array with the times of the trajectory in
"YYYY-MM-DD" - for now restricted to be time.}

\item{xName}{- character string with name of the xcoord in the ERDDAP dataset (default "longitude")}

\item{yName}{- character string with name of the ycoord in the ERDDAP dataset (default "latitude")}

\item{zName}{- character string with name of the zcoord in the ERDDAP dataset (default "altitude")}

\item{tName}{- character string with name of the tcoord in the ERDDAP dataset (default "time")}

\item{urlbase}{- base URL of the ERDDAP server being accessed - default "http://upwell.pfeg.noaa.gov/erddap"}

\item{verbose}{- logical variable (default FALSE) if the the URL request should be verbose}
}
\value{
structure with data and dimensions:
\itemize{
  \item extract$data - the data array dimensioned (lon,lat,time)
  \item extract$varname - the name of the parameter extracted
  \item extract$datasetname - ERDDAP dataset name
  \item extract$longitude - the longitudes on some scale as request
  \item extract$latitude - the latitudes always going south to north
  \item extract$time - the times of the extracts
  }
}
\description{
\code{rxtracto_3D} uses the R program rerddap to extract environmental data
from an ERDDAP server in an (x,y,z, time) bounding box.
The same call could be made directly in rerddap,
but function is maintained as it is used in the polygon routine.
}
\examples{
urlbase <- 'http://upwell.pfeg.noaa.gov/erddap'
dataInfo <- rerddap::info('erdPHssta8day')
parameter <- 'sst'
xcoord <- c(230,235)
ycoord <- c(40,45)
tcoord <- c('2006-01-15','2006-01-20')
zcoord <- 0.
extract <- rxtracto_3D(dataInfo, parameter, xcoord=xcoord, ycoord=ycoord, tcoord=tcoord, zcoord=zcoord)
2-D example getting bathymetry
dataInfo <- rerddap::info('etopo360')
parameter <- 'altitude'
extract <- xtracto_3D(dataInfo, parameter, xcoord=xcoord, ycoord=ycoord)
Example where grid is not latitude-longitude
dataInfo <- rerddap::info('glos_tds_5912_ca66_3f41')
parameter = 'temp'
xName <- 'nx'
yName <- 'ny'
zName <- 'nsigma'
xcoord <- c(10,11)
ycoord <- c(10,11)
zcoord <- c(1,1)
tcoord <- c('2016-09-02','2016-09-03')
extract <- rxtracto_3D(dataInfo, parameter, xcoord=xcoord, ycoord=ycoord, zcoord=zcoord, tcoord=tcoord, xName=xName, yName=yName, zName=zName)
Dataset that has depth also
3 months of subsurface temperature at 70m depth from SODA 2.2.4
dataInfo <- rerddap::info('hawaii_d90f_20ee_c4cb')
parameter = 'temp'
xName <- 'longitude'
yName <- 'latitude'
zName <- 'depth'
xcoord <- c(230.25,250.25)
ycoord <- c(30.25,43.25)
zcoord <- c(70.02,70.02)
tcoord <- c('2010-10-15','2010-12-15')
extract <- rxtracto_3D(dataInfo, parameter, xcoord=xcoord, ycoord=ycoord, zcoord=zcoord, tcoord=tcoord, xName=xName, yName=yName, zName=zName)

}

