% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_info}
\alias{safe_info}
\title{Wrap rerddap::info() so that it will fail gracefully
\code{safe_info} uses tryCatch() to deal with unavailable response}
\usage{
safe_info(dataset_id, url = "https://upwell.pfeg.noaa.gov/erddap/")
}
\arguments{
\item{dataset_id}{ERDDAP dataset_id}

\item{url}{base ERDDAP URL}
}
\value{
If success,  dataset information,  otherwise message and end
}
\description{
Wrap rerddap::info() so that it will fail gracefully
\code{safe_info} uses tryCatch() to deal with unavailable response
}
\examples{
dataInfo <- safe_info('etopo360')

}
