% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swchl.R
\docType{data}
\name{swchl}
\alias{swchl}
\title{swchl Data}
\format{An object of class \code{list} (inherits from \code{rxtractoTrack}) of length 13.}
\usage{
swchl
}
\description{
pre-Download of Pacific West Coast SST fro use in `plotTrack()` example
can run within CRAN Time limits
}
\details{
obtained using the `rxtracto()` command
tagData <- Marlintag38606
xpos <- tagData$lon[1:20]
ypos <- tagData$lat[1:20]
tpos <- tagData$date[1:20]
tpos <- tagData$date[1:20]
zpos <- rep(0., length(xpos))
swchlInfo <- rerddap::info('erdSWchla8day')
swchl <- rxtracto(swchlInfo, parameter = 'chlorophyll', xcoord = xpos,
                  ycoord = ypos, tcoord = tpos, zcoord = zpos, xlen = .2, ylen = .2)
}
\keyword{datasets}
